moved tiered stuff to child classes
